Switched to gradle and created different source sets
