+++ .travis.yml
@@ -1,0 +1,9 @@
+language: java
+jdk: oraclejdk8
+before_install: chmod +x gradlew
+script: ./gradlew
+notifications:
+  email: false
+env:
+  global:
+    TERM=dumb
