+++ api/me/flibio/updatifier/Updatifier.java
@@ -1,0 +1,58 @@
+/**
+ * This file is part of Updatifier, licensed under the MIT License (MIT).
+ * <p>
+ * Copyright (c) Flibio <http://github.com/Flibio>
+ * Copyright (c) contributors
+ * <p>
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * <p>
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * <p>
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package me.flibio.updatifier;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface Updatifier {
+
+    /**
+     * Gets the owner of the GitHub repository that the releases are hosted on.
+     *
+     * @return The owner of the GitHub repository that the releases are hosted on.
+     */
+    String repoOwner();
+
+    /**
+     * Gets the name of the GitHub repository that the releases are hosted on.
+     *
+     * @return The name of the GitHub repository that the releases are hosted on.
+     */
+    String repoName();
+
+    /**
+     * Gets the current version of the plugin. Must be exactly the same as the
+     * corresponding tag on GitHub.
+     *
+     * @return The current version of the plugin
+     */
+    String version();
+
+}
