+++ build.gradle
@@ -1,0 +1,66 @@
+plugins {
+  id 'java'
+  id 'com.github.hierynomus.license' version '0.12.1'
+}
+
+defaultTasks 'clean', 'build'
+
+sourceCompatibility = '1.8'
+targetCompatibility = '1.8'
+
+archivesBaseName = project.name.toLowerCase()
+version = "${project.version}-r${project.apiVersion.split("-")[0]}"
+
+repositories {
+  mavenCentral()
+  jcenter()
+  maven {
+    name 'Sponge maven repo'
+    url 'http://repo.spongepowered.org/maven'
+  }
+}
+
+dependencies {
+  compile "org.spongepowered:spongeapi:${project.apiVersion}"
+}
+
+sourceSets {
+  api {
+    java.srcDir 'api'
+    compileClasspath += configurations.compile
+  }
+  main {
+    java.srcDir 'src'
+    compileClasspath += api.output
+  }
+}
+
+task apiJar(type: Jar) {
+  classifier = 'api'
+  from sourceSets.api.output
+}
+
+jar {
+  from sourceSets.api.output
+  from sourceSets.main.output
+}
+
+license {
+  ext.name = project.name
+  ext.organization = project.organization
+  ext.url = project.url
+  header rootProject.file('HEADER.txt')
+  ignoreFailures true
+  strictCheck true
+  include '**/*.java'
+}
+
+build.dependsOn apiJar
+
+task wrapper(type: Wrapper) {
+  gradleVersion = '2.10'
+}
+
+tasks.withType(Javadoc) {
+  options.addStringOption('Xdoclint:none', '-quiet')
+}
