+++ gradle.properties
@@ -1,0 +1,6 @@
+name=Updatifier
+group=me.flibio
+url=http://github.com/Flibio
+organization=Flibio
+version=1.0.0-SNAPSHOT
+apiVersion=3.1.0-SNAPSHOT
